+++ web/package.json
@@ -11,7 +11,7 @@
     "@riophae/vue-treeselect": "^0.4.0",
     "@sentry/tracing": "^6.2.1",
     "@sentry/vue": "^6.2.1",
-    "@vue/composition-api": "^1.0.0-beta.25",
+    "@vue/composition-api": "^1.1.4",
     "axios": "^0.21.1",
     "axios-cache-adapter": "^2.7.3",
     "citation-js": "^0.5.0-alpha.4",
