+++ web/src/components/Wrappers/BinnedSummaryWrapper.vue
