+++ web/src/views/IndividualResults/IndividualTitle.vue
